topohedral_linalg

Struct IndexedMatrixViewMut

pub struct IndexedMatrixViewMut<'a, Mat>
where
    Mat: Shape + Index<(usize, usize)> + IndexMut<(usize, usize)>,
{ /* private fields */ }

Mutable subview of a matrix selected by unique row and column index lists.

Implementations

impl<'a, Mat> IndexedMatrixViewMut<'a, Mat>
where Mat: Shape + Index<(usize, usize)> + IndexMut<(usize, usize)>, Mat::Output: Copy + Sized,

pub fn iter(&'a self) -> IndexedMatrixViewMutIter<'a, Mat>

Returns a shared column-major iterator over the elements of this mutable view.

pub fn iter_mut(&'a mut self) -> IndexedMatrixViewMutIterMut<'a, Mat>

Returns a mutable column-major iterator over the elements of this mutable view.

pub fn copy_from<Rhs>(&mut self, rhs: Rhs)
where Rhs: MatrixExpr<ScalarType = Mat::Output>,

Copies entries from rhs into this mutable view.

Panics

Panics when rhs dimensions do not match this view’s dimensions.

pub fn to_dmatrix(&self) -> DMatrix<Mat::Output>

Copies the view contents into a new heap-allocated DMatrix.