topohedral_linalg

Trait MatrixElementDisplay

pub trait MatrixElementDisplay {
    // Required method
    fn fmt_matrix_element(&self, f: &mut Formatter<'_>) -> Result;
}

Controls how a scalar is rendered by matrix fmt::Display implementations.

Implement this trait for custom scalar types that should use the crate’s aligned matrix display format.

Required Methods

fn fmt_matrix_element(&self, f: &mut Formatter<'_>) -> Result

Writes one matrix element to the supplied formatter.