topohedral_linalg

Trait FloatTransformOps

pub trait FloatTransformOps: TransformOps
where
    Self::ScalarType: Float,
{

    // Provided methods
    fn abs(&mut self) { ... }
    fn absed(&self) -> Self
       where Self: Clone { ... }
    fn into_absed(self) -> Self { ... }
    fn abs_sub(&mut self, other: Self::ScalarType) { ... }
    fn abs_subed(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_abs_subed(self, other: Self::ScalarType) -> Self { ... }
    fn acos(&mut self) { ... }
    fn acosed(&self) -> Self
       where Self: Clone { ... }
    fn into_acosed(self) -> Self { ... }
    fn acosh(&mut self) { ... }
    fn acoshed(&self) -> Self
       where Self: Clone { ... }
    fn into_acoshed(self) -> Self { ... }
    fn asin(&mut self) { ... }
    fn asined(&self) -> Self
       where Self: Clone { ... }
    fn into_asined(self) -> Self { ... }
    fn asinh(&mut self) { ... }
    fn asinhed(&self) -> Self
       where Self: Clone { ... }
    fn into_asinhed(self) -> Self { ... }
    fn atan(&mut self) { ... }
    fn ataned(&self) -> Self
       where Self: Clone { ... }
    fn into_ataned(self) -> Self { ... }
    fn atan2(&mut self, other: Self::ScalarType) { ... }
    fn atan2ed(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_atan2ed(self, other: Self::ScalarType) -> Self { ... }
    fn atanh(&mut self) { ... }
    fn atanhed(&self) -> Self
       where Self: Clone { ... }
    fn into_atanhed(self) -> Self { ... }
    fn cbrt(&mut self) { ... }
    fn cbrted(&self) -> Self
       where Self: Clone { ... }
    fn into_cbrted(self) -> Self { ... }
    fn ceil(&mut self) { ... }
    fn ceiled(&self) -> Self
       where Self: Clone { ... }
    fn into_ceiled(self) -> Self { ... }
    fn clamp(&mut self, min: Self::ScalarType, max: Self::ScalarType) { ... }
    fn clamped(&self, min: Self::ScalarType, max: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_clamped(self, min: Self::ScalarType, max: Self::ScalarType) -> Self { ... }
    fn copysign(&mut self, sign: Self::ScalarType) { ... }
    fn copysigned(&self, sign: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_copysigned(self, sign: Self::ScalarType) -> Self { ... }
    fn cos(&mut self) { ... }
    fn cosed(&self) -> Self
       where Self: Clone { ... }
    fn into_cosed(self) -> Self { ... }
    fn cosh(&mut self) { ... }
    fn coshed(&self) -> Self
       where Self: Clone { ... }
    fn into_coshed(self) -> Self { ... }
    fn div_euclid(&mut self, rhs: Self::ScalarType) { ... }
    fn div_euclided(&self, rhs: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_div_euclided(self, rhs: Self::ScalarType) -> Self { ... }
    fn exp(&mut self) { ... }
    fn exped(&self) -> Self
       where Self: Clone { ... }
    fn into_exped(self) -> Self { ... }
    fn exp2(&mut self) { ... }
    fn exp2ed(&self) -> Self
       where Self: Clone { ... }
    fn into_exp2ed(self) -> Self { ... }
    fn exp_m1(&mut self) { ... }
    fn exp_m1ed(&self) -> Self
       where Self: Clone { ... }
    fn into_exp_m1ed(self) -> Self { ... }
    fn floor(&mut self) { ... }
    fn floored(&self) -> Self
       where Self: Clone { ... }
    fn into_floored(self) -> Self { ... }
    fn fract(&mut self) { ... }
    fn fracted(&self) -> Self
       where Self: Clone { ... }
    fn into_fracted(self) -> Self { ... }
    fn hypot(&mut self, other: Self::ScalarType) { ... }
    fn hypoted(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_hypoted(self, other: Self::ScalarType) -> Self { ... }
    fn ln(&mut self) { ... }
    fn lned(&self) -> Self
       where Self: Clone { ... }
    fn into_lned(self) -> Self { ... }
    fn ln_1p(&mut self) { ... }
    fn ln_1ped(&self) -> Self
       where Self: Clone { ... }
    fn into_ln_1ped(self) -> Self { ... }
    fn log(&mut self, base: Self::ScalarType) { ... }
    fn loged(&self, base: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_loged(self, base: Self::ScalarType) -> Self { ... }
    fn log10(&mut self) { ... }
    fn log10ed(&self) -> Self
       where Self: Clone { ... }
    fn into_log10ed(self) -> Self { ... }
    fn log2(&mut self) { ... }
    fn log2ed(&self) -> Self
       where Self: Clone { ... }
    fn into_log2ed(self) -> Self { ... }
    fn max(&mut self, other: Self::ScalarType) { ... }
    fn maxed(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_maxed(self, other: Self::ScalarType) -> Self { ... }
    fn midpoint(&mut self, other: Self::ScalarType) { ... }
    fn midpointed(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_midpointed(self, other: Self::ScalarType) -> Self { ... }
    fn min(&mut self, other: Self::ScalarType) { ... }
    fn mined(&self, other: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_mined(self, other: Self::ScalarType) -> Self { ... }
    fn mul_add(&mut self, a: Self::ScalarType, b: Self::ScalarType) { ... }
    fn mul_added(&self, a: Self::ScalarType, b: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_mul_added(self, a: Self::ScalarType, b: Self::ScalarType) -> Self { ... }
    fn next_down(&mut self) { ... }
    fn next_downed(&self) -> Self
       where Self: Clone { ... }
    fn into_next_downed(self) -> Self { ... }
    fn next_up(&mut self) { ... }
    fn next_uped(&self) -> Self
       where Self: Clone { ... }
    fn into_next_uped(self) -> Self { ... }
    fn powf(&mut self, exp: Self::ScalarType) { ... }
    fn powfed(&self, exp: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_powfed(self, exp: Self::ScalarType) -> Self { ... }
    fn powi(&mut self, exp: i32) { ... }
    fn powied(&self, exp: i32) -> Self
       where Self: Clone { ... }
    fn into_powied(self, exp: i32) -> Self { ... }
    fn recip(&mut self) { ... }
    fn reciped(&self) -> Self
       where Self: Clone { ... }
    fn into_reciped(self) -> Self { ... }
    fn rem_euclid(&mut self, rhs: Self::ScalarType) { ... }
    fn rem_euclided(&self, rhs: Self::ScalarType) -> Self
       where Self: Clone { ... }
    fn into_rem_euclided(self, rhs: Self::ScalarType) -> Self { ... }
    fn round(&mut self) { ... }
    fn rounded(&self) -> Self
       where Self: Clone { ... }
    fn into_rounded(self) -> Self { ... }
    fn round_ties_even(&mut self) { ... }
    fn round_ties_evened(&self) -> Self
       where Self: Clone { ... }
    fn into_round_ties_evened(self) -> Self { ... }
    fn signum(&mut self) { ... }
    fn signumed(&self) -> Self
       where Self: Clone { ... }
    fn into_signumed(self) -> Self { ... }
    fn sin(&mut self) { ... }
    fn sined(&self) -> Self
       where Self: Clone { ... }
    fn into_sined(self) -> Self { ... }
    fn sinh(&mut self) { ... }
    fn sinhed(&self) -> Self
       where Self: Clone { ... }
    fn into_sinhed(self) -> Self { ... }
    fn sqrt(&mut self) { ... }
    fn sqrted(&self) -> Self
       where Self: Clone { ... }
    fn into_sqrted(self) -> Self { ... }
    fn tan(&mut self) { ... }
    fn taned(&self) -> Self
       where Self: Clone { ... }
    fn into_taned(self) -> Self { ... }
    fn tanh(&mut self) { ... }
    fn tanhed(&self) -> Self
       where Self: Clone { ... }
    fn into_tanhed(self) -> Self { ... }
    fn to_degrees(&mut self) { ... }
    fn to_degreesed(&self) -> Self
       where Self: Clone { ... }
    fn into_to_degreesed(self) -> Self { ... }
    fn to_radians(&mut self) { ... }
    fn to_radiansed(&self) -> Self
       where Self: Clone { ... }
    fn into_to_radiansed(self) -> Self { ... }
    fn trunc(&mut self) { ... }
    fn trunced(&self) -> Self
       where Self: Clone { ... }
    fn into_trunced(self) -> Self { ... }
    fn pos(&mut self)
       where Self::ScalarType: Zero { ... }
    fn posed(&self) -> Self
       where Self: Clone,
             Self::ScalarType: Zero { ... }
    fn into_posed(self) -> Self
       where Self::ScalarType: Zero { ... }
    fn neg(&mut self)
       where Self::ScalarType: Zero { ... }
    fn neged(&self) -> Self
       where Self: Clone,
             Self::ScalarType: Zero { ... }
    fn into_neged(self) -> Self
       where Self::ScalarType: Zero { ... }

}

Extends TransformOps with element-wise versions of every Float method, providing in-place, cloning (*ed), and consuming (into_*ed) variants for each operation.

Provided Methods

fn abs(&mut self)

fn absed(&self) -> Self
where Self: Clone,

fn into_absed(self) -> Self

fn abs_sub(&mut self, other: Self::ScalarType)

fn abs_subed(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_abs_subed(self, other: Self::ScalarType) -> Self

fn acos(&mut self)

fn acosed(&self) -> Self
where Self: Clone,

fn into_acosed(self) -> Self

fn acosh(&mut self)

fn acoshed(&self) -> Self
where Self: Clone,

fn into_acoshed(self) -> Self

fn asin(&mut self)

fn asined(&self) -> Self
where Self: Clone,

fn into_asined(self) -> Self

fn asinh(&mut self)

fn asinhed(&self) -> Self
where Self: Clone,

fn into_asinhed(self) -> Self

fn atan(&mut self)

fn ataned(&self) -> Self
where Self: Clone,

fn into_ataned(self) -> Self

fn atan2(&mut self, other: Self::ScalarType)

fn atan2ed(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_atan2ed(self, other: Self::ScalarType) -> Self

fn atanh(&mut self)

fn atanhed(&self) -> Self
where Self: Clone,

fn into_atanhed(self) -> Self

fn cbrt(&mut self)

fn cbrted(&self) -> Self
where Self: Clone,

fn into_cbrted(self) -> Self

fn ceil(&mut self)

fn ceiled(&self) -> Self
where Self: Clone,

fn into_ceiled(self) -> Self

fn clamp(&mut self, min: Self::ScalarType, max: Self::ScalarType)

fn clamped(&self, min: Self::ScalarType, max: Self::ScalarType) -> Self
where Self: Clone,

fn into_clamped(self, min: Self::ScalarType, max: Self::ScalarType) -> Self

fn copysign(&mut self, sign: Self::ScalarType)

fn copysigned(&self, sign: Self::ScalarType) -> Self
where Self: Clone,

fn into_copysigned(self, sign: Self::ScalarType) -> Self

fn cos(&mut self)

fn cosed(&self) -> Self
where Self: Clone,

fn into_cosed(self) -> Self

fn cosh(&mut self)

fn coshed(&self) -> Self
where Self: Clone,

fn into_coshed(self) -> Self

fn div_euclid(&mut self, rhs: Self::ScalarType)

fn div_euclided(&self, rhs: Self::ScalarType) -> Self
where Self: Clone,

fn into_div_euclided(self, rhs: Self::ScalarType) -> Self

fn exp(&mut self)

fn exped(&self) -> Self
where Self: Clone,

fn into_exped(self) -> Self

fn exp2(&mut self)

fn exp2ed(&self) -> Self
where Self: Clone,

fn into_exp2ed(self) -> Self

fn exp_m1(&mut self)

fn exp_m1ed(&self) -> Self
where Self: Clone,

fn into_exp_m1ed(self) -> Self

fn floor(&mut self)

fn floored(&self) -> Self
where Self: Clone,

fn into_floored(self) -> Self

fn fract(&mut self)

fn fracted(&self) -> Self
where Self: Clone,

fn into_fracted(self) -> Self

fn hypot(&mut self, other: Self::ScalarType)

fn hypoted(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_hypoted(self, other: Self::ScalarType) -> Self

fn ln(&mut self)

fn lned(&self) -> Self
where Self: Clone,

fn into_lned(self) -> Self

fn ln_1p(&mut self)

fn ln_1ped(&self) -> Self
where Self: Clone,

fn into_ln_1ped(self) -> Self

fn log(&mut self, base: Self::ScalarType)

fn loged(&self, base: Self::ScalarType) -> Self
where Self: Clone,

fn into_loged(self, base: Self::ScalarType) -> Self

fn log10(&mut self)

fn log10ed(&self) -> Self
where Self: Clone,

fn into_log10ed(self) -> Self

fn log2(&mut self)

fn log2ed(&self) -> Self
where Self: Clone,

fn into_log2ed(self) -> Self

fn max(&mut self, other: Self::ScalarType)

fn maxed(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_maxed(self, other: Self::ScalarType) -> Self

fn midpoint(&mut self, other: Self::ScalarType)

fn midpointed(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_midpointed(self, other: Self::ScalarType) -> Self

fn min(&mut self, other: Self::ScalarType)

fn mined(&self, other: Self::ScalarType) -> Self
where Self: Clone,

fn into_mined(self, other: Self::ScalarType) -> Self

fn mul_add(&mut self, a: Self::ScalarType, b: Self::ScalarType)

fn mul_added(&self, a: Self::ScalarType, b: Self::ScalarType) -> Self
where Self: Clone,

fn into_mul_added(self, a: Self::ScalarType, b: Self::ScalarType) -> Self

fn next_down(&mut self)

fn next_downed(&self) -> Self
where Self: Clone,

fn into_next_downed(self) -> Self

fn next_up(&mut self)

fn next_uped(&self) -> Self
where Self: Clone,

fn into_next_uped(self) -> Self

fn powf(&mut self, exp: Self::ScalarType)

fn powfed(&self, exp: Self::ScalarType) -> Self
where Self: Clone,

fn into_powfed(self, exp: Self::ScalarType) -> Self

fn powi(&mut self, exp: i32)

fn powied(&self, exp: i32) -> Self
where Self: Clone,

fn into_powied(self, exp: i32) -> Self

fn recip(&mut self)

fn reciped(&self) -> Self
where Self: Clone,

fn into_reciped(self) -> Self

fn rem_euclid(&mut self, rhs: Self::ScalarType)

fn rem_euclided(&self, rhs: Self::ScalarType) -> Self
where Self: Clone,

fn into_rem_euclided(self, rhs: Self::ScalarType) -> Self

fn round(&mut self)

fn rounded(&self) -> Self
where Self: Clone,

fn into_rounded(self) -> Self

fn round_ties_even(&mut self)

fn round_ties_evened(&self) -> Self
where Self: Clone,

fn into_round_ties_evened(self) -> Self

fn signum(&mut self)

fn signumed(&self) -> Self
where Self: Clone,

fn into_signumed(self) -> Self

fn sin(&mut self)

fn sined(&self) -> Self
where Self: Clone,

fn into_sined(self) -> Self

fn sinh(&mut self)

fn sinhed(&self) -> Self
where Self: Clone,

fn into_sinhed(self) -> Self

fn sqrt(&mut self)

fn sqrted(&self) -> Self
where Self: Clone,

fn into_sqrted(self) -> Self

fn tan(&mut self)

fn taned(&self) -> Self
where Self: Clone,

fn into_taned(self) -> Self

fn tanh(&mut self)

fn tanhed(&self) -> Self
where Self: Clone,

fn into_tanhed(self) -> Self

fn to_degrees(&mut self)

fn to_degreesed(&self) -> Self
where Self: Clone,

fn into_to_degreesed(self) -> Self

fn to_radians(&mut self)

fn to_radiansed(&self) -> Self
where Self: Clone,

fn into_to_radiansed(self) -> Self

fn trunc(&mut self)

fn trunced(&self) -> Self
where Self: Clone,

fn into_trunced(self) -> Self

fn pos(&mut self)
where Self::ScalarType: Zero,

Clamps every element to the positive part in-place (negative values become zero).

fn posed(&self) -> Self
where Self: Clone, Self::ScalarType: Zero,

Returns a cloned copy of self with every element clamped to its positive part.

fn into_posed(self) -> Self
where Self::ScalarType: Zero,

Consumes self, clamps every element to its positive part, and returns it.

fn neg(&mut self)
where Self::ScalarType: Zero,

Clamps every element to the non-positive part in-place (positive values become zero).

fn neged(&self) -> Self
where Self: Clone, Self::ScalarType: Zero,

Returns a cloned copy of self with every element clamped to its non-positive part.

fn into_neged(self) -> Self
where Self::ScalarType: Zero,

Consumes self, clamps every element to its non-positive part, and returns it.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.